+++ app/Livewire/Web/Chat/ChatBox.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Web\Chat;
 
 use App\Models\ChatRoom;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class ChatBox extends Component
@@ -10,20 +11,38 @@ class ChatBox extends Component
 
     public ChatRoom $chatRoom;
     public $messages;
+    public $entryTime;
 
-     protected $listeners = ['novaMensagem' => 'atualizarMensagens', 'messageSentLocally' => 'atualizarMensagens'];
+    protected $listeners = ['novaMensagem' => 'atualizarMensagens', 'messageSentLocally' => 'atualizarMensagens'];
 
     public function mount(ChatRoom $chatRoom)
     {
         $this->chatRoom = $chatRoom;
-        $this->messages = $chatRoom->messages()->with('user')->get();
+
+        $user = Auth::user();
+        if ($user) {
+            $pivotData = $user->chatRooms()
+                ->where('chat_rooms.id', $this->chatRoom->id)
+                ->first()?->pivot;
+
+            if ($pivotData) {
+                $this->entryTime = $pivotData->entered_room;
+            }
+        }
+
+        $this->messages = $this->chatRoom->messages()
+            ->where('send_at', '>=', $this->entryTime)
+            ->with('user')
+            ->get();
     }
+
 
     public function atualizarMensagens()
     {
-        $this->messages = $this->chatRoom->messages()->with('user')->get();
-
-        //$this->dispatch('scrollChatToBottom');
+        $this->messages = $this->chatRoom->messages()
+            ->where('send_at', '>=', $this->entryTime)
+            ->with('user')
+            ->get();
     }
 
     public function render()